requirements.txt
openai
aiogram
psycopg2-binary
python-dotenv